refactor: extract getIconByLabel into factory function.
